requirements.txt
alembic==1.13.3
babel==2.16.0
BareNecessities==0.2.8
blinker==1.8.2
click==8.1.7
colorama==0.4.6
coverage==7.6.10
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Admin==1.6.1
flask-babel==4.0.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mail==2.1.0
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
PyJWT==2.9.0
PyMySQL==1.1.1
pytest==8.3.4
python-dotenv==1.0.1
pytz==2024.2
ruff==0.9.6
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==3.0.4
WTForms==3.1.2
